PLAYE_Desktop_AI_stage3/backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
pydantic-settings==2.1.0

# Queue
celery==5.3.6
redis==5.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# PyTorch (CPU fallback — замени на GPU-версию при наличии CUDA)
torch==2.1.0
torchvision==0.16.0
onnxruntime==1.16.0

# Image processing
opencv-python-headless==4.8.1.78
numpy==1.26.2
Pillow==10.1.0

# AI models
ultralytics==8.0.196
retinaface==0.0.1
basicsr==1.4.2
gfpgan==1.3.8
realesrgan==0.3.0

# Utils
requests==2.31.0
tqdm==4.66.1

# Enterprise auth
bcrypt==4.1.2
python-jose==3.3.0

# ffmpeg для видео
ffmpeg-python==0.2.0

# Monitoring
prometheus-fastapi-instrumentator==6.1.0
